+++ routes/web.php
@@ -6,3 +6,6 @@ use Illuminate\Support\Facades\Route;
 //Main Page
 Route::get('/', [YoutubeChannelController::class, 'index'])
     ->name('index');
+
+Route::get('/feeds/{youtubeChannel:youtube_id}', [YoutubeChannelController::class, 'feed'])
+    ->name('feeds.show');